--- src/app/pages/consultas-ordenesdepago/components/layout/layout.component.ts
@@ -1,18 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'ngx-layout',
-  templateUrl: './layout.component.html',
-  styleUrls: ['./layout.component.scss']
-})
-export class LayoutComponent implements OnInit {
-
-
-  titulo: any;
-  constructor() {
-    this.titulo = 'REGISTRO ORDENES DE PAGO';
-  }
-  ngOnInit() {
-  }
-
-}
