--- src/app/pages/consultas-ordenesdepago/ordenespago-routing.module.ts
@@ -1,22 +1,0 @@
-import { NgModule } from '@angular/core';
-import { Routes, RouterModule } from '@angular/router';
-import { LayoutComponent } from './components/layout/layout.component';
-
-const routes: Routes = [
-  {
-    path: '',
-    component: LayoutComponent,
-    children: [
-      {
-        path: 'consultaordenespago',
-        loadChildren: () => import('./components/layout/layout.component')
-        .then(m => m.LayoutComponent),
-      },
-    ],
-}];
-
-@NgModule({
-  imports: [RouterModule.forChild(routes)],
-  exports: [RouterModule]
-})
-export class OrdenespagoRoutingModule { }
