+++ src/app/pages/ordenesdepago/components/crear-ordenpago/crear-ordenpago.component.spec.ts
@@ -1,20 +1,20 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
-import { LayoutComponent } from './layout.component';
+import { CrearOrdenpagoComponent } from './crear-ordenpago.component';
 
-describe('LayoutComponent', () => {
-  let component: LayoutComponent;
-  let fixture: ComponentFixture<LayoutComponent>;
+describe('CrearOrdenpagoComponent', () => {
+  let component: CrearOrdenpagoComponent;
+  let fixture: ComponentFixture<CrearOrdenpagoComponent>;
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-      declarations: [ LayoutComponent ]
+      declarations: [ CrearOrdenpagoComponent ]
     })
     .compileComponents();
   }));
 
   beforeEach(() => {
-    fixture = TestBed.createComponent(LayoutComponent);
+    fixture = TestBed.createComponent(CrearOrdenpagoComponent);
     component = fixture.componentInstance;
     fixture.detectChanges();
   });
