+++ src/app/pages/ordenesdepago/components/crear-ordenpago/crear-ordenpago.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'ngx-crear-ordenpago',
+  templateUrl: './crear-ordenpago.component.html',
+  styleUrls: ['./crear-ordenpago.component.scss']
+})
+export class CrearOrdenpagoComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit() {
+  }
+
+}
