+++ src/app/pages/ordenesdepago/components/table-listaordenes/table-listaordenes.component.html
@@ -1,4 +1,4 @@
-<div class="container">
+<div class="container bg-white p-5 border rounded">
     <div class="row">
         <div class="col py-2">
             <h5 class="card-title text-center text-primary">REGISTRO DE ORDENES DE PAGO</h5>
