+++ src/app/pages/ordenesdepago/interfaces/interfaces.ts
